+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre release
 
 ### Enhancements
+* IOO-420 - International Online Offer triage question 4 and 5 allowing user to input business spend and hiring
 * IOO-418 - International Online Offer triage question 3 allowing user to select business location
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
@@ -15,6 +16,7 @@
 
 ### Bugs fixed
 * KLS-425 - Update broken links for export opportunities
+* IOO-418 - International Online Offer triage question 3 allowing user to select business location
 
 ## [2.26.0](https://github.com/uktrade/great-cms/releases/tag/2.26.0)
 ### Enhancements