+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * KLS-342 - Fixes on export academy landing page - squash migrtions, add intro_text field and update spacing
 * KLS-427 - Export academy event reminder notification
 * KLS-355 - Export Academy registration journey
+* KLS-437 - Change DIT references to DBT
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/2.26.0...2.27.0)