+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * KLS-427 - Export academy event reminder notification
 * KLS-355 - Export Academy registration journey
 * KLS-363 - Created Microsite template
+* KLS-437 - Change DIT references to DBT
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/2.26.0...2.27.0)