+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * KLS-355 - Export Academy registration journey
 * KLS-363 - Created Microsite template
 * KLS-437 - Change DIT references to DBT
+* KLS-407 - Add format filters to the Export Academy event listings page
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/2.26.0...2.27.0)
@@ -18,7 +19,6 @@
 * IOO-402 - International Online Offer triage question 1 allowing user to select business sector
 * KLS-342 - Update export academy landing page to be in line with design
 * KLS-353 - Export Academy booking and cancellation confirmation emails
-* KLS-407 - Add format filters to the Export Academy event listings page
 
 ### Bugs fixed
 * KLS-425 - Update broken links for export opportunities