+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Pre release
 
 ### Enhancements
+* KLS-426 - Add event types to filterset on UKEA's listings page
+
+## [2.28.0](https://github.com/uktrade/great-cms/releases/tag/2.28.0)
+[Full Changelog](https://github.com/uktrade/great-cms/compare/2.27.0...2.28.0)
+### Enhancements
 * IOO-420 - International Online Offer triage question 4 and 5 allowing user to input business spend and hiring
 * IOO-418 - International Online Offer triage question 3 allowing user to select business location
 * KLS-342 - Fixes on export academy landing page - squash migrtions, add intro_text field and update spacing
@@ -12,7 +17,6 @@
 * KLS-437 - Change DIT references to DBT
 * KLS-407 - Add format filters to the Export Academy event listings page
 * KLS-270 - Add webinar duration to event listing template
-* KLS-426 - Add event types to filterset on UKEA's listings page
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/2.26.0...2.27.0)